# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DeadzoneType = (typeof DeadzoneType)[keyof typeof DeadzoneType];

export const DeadzoneType = {
  cellular: "cellular",
  wifi: "wifi",
  gps: "gps",
  satellite: "satellite",
} as const;

export type DeadzoneSeverity =
  (typeof DeadzoneSeverity)[keyof typeof DeadzoneSeverity];

export const DeadzoneSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  total: "total",
} as const;

export interface Deadzone {
  id: number;
  title: string;
  description: string;
  type: DeadzoneType;
  severity: DeadzoneSeverity;
  /** Measured signal strength in dBm (e.g. -110) */
  signalStrength?: number | null;
  latitude: number;
  longitude: number;
  carrier: string;
  reporter: string;
  confirmations: number;
  createdAt: string;
}

export type DeadzoneDetail = Deadzone & {
  address?: string | null;
};

export interface CreateDeadzone {
  /** @minLength 1 */
  title: string;
  description: string;
  type: DeadzoneType;
  severity: DeadzoneSeverity;
  /** Measured signal strength in dBm (e.g. -110) */
  signalStrength?: number;
  latitude: number;
  longitude: number;
  carrier: string;
  reporter: string;
  address?: string;
}

export type StatsSummaryByTypeItem = {
  type: DeadzoneType;
  count: number;
};

export type StatsSummaryBySeverityItem = {
  severity: DeadzoneSeverity;
  count: number;
};

export interface StatsSummary {
  totalReports: number;
  totalConfirmations: number;
  uniqueCarriers: number;
  activeCities: number;
  avgSignalStrength?: number | null;
  byType: StatsSummaryByTypeItem[];
  bySeverity: StatsSummaryBySeverityItem[];
}

export interface Hotspot {
  label: string;
  latitude: number;
  longitude: number;
  count: number;
  topType: DeadzoneType;
}

export interface CarrierStat {
  carrier: string;
  count: number;
  avgConfirmations: number;
  avgSignalStrength?: number | null;
}

export interface Cluster {
  id: number;
  centroidLat: number;
  centroidLng: number;
  count: number;
  radiusKm: number;
  /** Weighted confidence score 0-1 based on recency, density and confirmations */
  confidence: number;
  avgSignalStrength?: number | null;
  topSeverity: DeadzoneSeverity;
  topCarrier: string;
  lastUpdated: string;
  reportIds?: number[];
}

export type ListDeadzonesParams = {
  type?: DeadzoneType;
  severity?: DeadzoneSeverity;
  carrier?: string;
};

export type ListRecentDeadzonesParams = {
  limit?: number;
};

export type GetHotspotsParams = {
  limit?: number;
};

export type GetClustersParams = {
  /**
   * DBSCAN neighborhood radius in kilometers
   */
  epsKm?: number;
  /**
   * DBSCAN minimum points to form a dense cluster
   */
  minPts?: number;
};