# lib/api-zod/src/generated/types/carrierStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */

export interface CarrierStat {
  carrier: string;
  count: number;
  avgConfirmations: number;
  avgSignalStrength?: number | null;
}