# lib/api-zod/src/generated/types/cluster.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */
import type { DeadzoneSeverity } from "./deadzoneSeverity";

export interface Cluster {
  id: number;
  centroidLat: number;
  centroidLng: number;
  count: number;
  radiusKm: number;
  /** Weighted confidence score 0-1 based on recency, density and confirmations */
  confidence: number;
  avgSignalStrength?: number | null;
  topSeverity: DeadzoneSeverity;
  topCarrier: string;
  lastUpdated: Date;
  reportIds?: number[];
}