# lib/api-zod/src/generated/types/createDeadzone.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */
import type { DeadzoneSeverity } from "./deadzoneSeverity";
import type { DeadzoneType } from "./deadzoneType";

export interface CreateDeadzone {
  /** @minLength 1 */
  title: string;
  description: string;
  type: DeadzoneType;
  severity: DeadzoneSeverity;
  /** Measured signal strength in dBm (e.g. -110) */
  signalStrength?: number;
  latitude: number;
  longitude: number;
  carrier: string;
  reporter: string;
  address?: string;
}