# lib/api-zod/src/generated/types/deadzone.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */
import type { DeadzoneSeverity } from "./deadzoneSeverity";
import type { DeadzoneType } from "./deadzoneType";

export interface Deadzone {
  id: number;
  title: string;
  description: string;
  type: DeadzoneType;
  severity: DeadzoneSeverity;
  /** Measured signal strength in dBm (e.g. -110) */
  signalStrength?: number | null;
  latitude: number;
  longitude: number;
  carrier: string;
  reporter: string;
  confirmations: number;
  createdAt: Date;
}