# lib/api-zod/src/generated/types/deadzoneDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */
import type { Deadzone } from "./deadzone";

export type DeadzoneDetail = Deadzone & {
  address?: string | null;
};