# lib/api-zod/src/generated/types/deadzoneSeverity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */

export type DeadzoneSeverity =
  (typeof DeadzoneSeverity)[keyof typeof DeadzoneSeverity];

export const DeadzoneSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  total: "total",
} as const;