# lib/api-zod/src/generated/types/deadzoneType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */

export type DeadzoneType = (typeof DeadzoneType)[keyof typeof DeadzoneType];

export const DeadzoneType = {
  cellular: "cellular",
  wifi: "wifi",
  gps: "gps",
  satellite: "satellite",
} as const;