# lib/api-zod/src/generated/types/hotspot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */
import type { DeadzoneType } from "./deadzoneType";

export interface Hotspot {
  label: string;
  latitude: number;
  longitude: number;
  count: number;
  topType: DeadzoneType;
}