# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dead Zone Tracker API (India)
 * OpenAPI spec version: 0.1.0
 */
import type { StatsSummaryBySeverityItem } from "./statsSummaryBySeverityItem";
import type { StatsSummaryByTypeItem } from "./statsSummaryByTypeItem";

export interface StatsSummary {
  totalReports: number;
  totalConfirmations: number;
  uniqueCarriers: number;
  activeCities: number;
  avgSignalStrength?: number | null;
  byType: StatsSummaryByTypeItem[];
  bySeverity: StatsSummaryBySeverityItem[];
}